extract http context to injected class
