+++ src/main/java/io/cryostat/agent/WebServer.java
@@ -37,15 +37,7 @@
  */
 package io.cryostat.agent;
 
-import java.io.ByteArrayOutputStream;
-import java.io.DataOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.lang.management.ManagementFactory;
-import java.lang.management.MemoryMXBean;
-import java.lang.management.OperatingSystemMXBean;
-import java.lang.management.RuntimeMXBean;
-import java.lang.management.ThreadMXBean;
 import java.net.InetSocketAddress;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
@@ -54,22 +46,14 @@ import java.security.SecureRandom;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Base64;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.Callable;
+import java.util.Set;
 import java.util.concurrent.ScheduledExecutorService;
 
-import io.cryostat.core.net.MBeanMetrics;
-import io.cryostat.core.net.MemoryMetrics;
-import io.cryostat.core.net.OperatingSystemMetrics;
-import io.cryostat.core.net.RuntimeMetrics;
-import io.cryostat.core.net.ThreadMetrics;
+import io.cryostat.agent.remote.RemoteContext;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sun.management.UnixOperatingSystemMXBean;
 import com.sun.net.httpserver.BasicAuthenticator;
 import com.sun.net.httpserver.Filter;
 import com.sun.net.httpserver.HttpContext;
@@ -77,7 +61,6 @@ import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import com.sun.net.httpserver.HttpServer;
 import dagger.Lazy;
-import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -86,6 +69,7 @@ class WebServer {
 
     private final Logger log = LoggerFactory.getLogger(getClass());
 
+    private final Lazy<Set<RemoteContext>> remoteContexts;
     private final ObjectMapper mapper;
     private final Lazy<CryostatClient> cryostat;
     private final ScheduledExecutorService executor;
@@ -96,12 +80,14 @@ class WebServer {
     private HttpServer http;
 
     WebServer(
+            Lazy<Set<RemoteContext>> remoteContexts,
             ObjectMapper mapper,
             Lazy<CryostatClient> cryostat,
             ScheduledExecutorService executor,
             String host,
             int port,
             Lazy<Registration> registration) {
+        this.remoteContexts = remoteContexts;
         this.mapper = mapper;
         this.cryostat = cryostat;
         this.executor = executor;
@@ -123,7 +109,6 @@ class WebServer {
         this.http.setExecutor(executor);
 
         List<HttpContext> ctxs = new ArrayList<>();
-
         ctxs.add(
                 this.http.createContext(
                         "/",
@@ -151,36 +136,10 @@ class WebServer {
                                 }
                             }
                         }));
-
-        ctxs.add(
-                this.http.createContext(
-                        "/mbean-metrics",
-                        new HttpHandler() {
-                            @Override
-                            public void handle(HttpExchange exchange) throws IOException {
-                                String mtd = exchange.getRequestMethod();
-                                switch (mtd) {
-                                    case "GET":
-                                        try {
-                                            MBeanMetrics metrics = getMbeanMetrics();
-                                            exchange.sendResponseHeaders(HttpStatus.SC_OK, 0);
-                                            try (OutputStream response =
-                                                    exchange.getResponseBody()) {
-                                                mapper.writeValue(response, metrics);
-                                            }
-                                        } catch (Exception e) {
-                                            log.error("mbean serialization failure", e);
-                                        } finally {
-                                            exchange.close();
-                                        }
-                                        break;
-                                    default:
-                                        exchange.sendResponseHeaders(HttpStatus.SC_NOT_FOUND, -1);
-                                        exchange.close();
-                                        break;
-                                }
-                            }
-                        }));
+        ctxs.addAll(
+                remoteContexts.get().stream()
+                        .map(rc -> this.http.createContext(rc.path(), rc::handle))
+                        .toList());
 
         ctxs.forEach(ctx -> ctx.setAuthenticator(new AgentAuthenticator()));
         ctxs.forEach(
@@ -234,141 +193,6 @@ class WebServer {
                     .thenAccept(i -> log.info("Defined credentials with id {}", i))
                     .thenRun(this.credentials::clear);
         }
-    }
-
-    private <T> void safeStore(String key, Map<String, Object> map, Callable<T> fn) {
-        try {
-            map.put(key, fn.call());
-        } catch (Exception e) {
-            log.warn("Call failed", e);
-        }
-    }
-
-    private MBeanMetrics getMbeanMetrics() {
-        // TODO refactor, extract into -core library?
-        RuntimeMXBean runtimeBean = ManagementFactory.getRuntimeMXBean();
-        Map<String, Object> runtimeAttrs = new HashMap<>();
-        safeStore("BootClassPath", runtimeAttrs, runtimeBean::getBootClassPath);
-        safeStore("ClassPath", runtimeAttrs, runtimeBean::getClassPath);
-        safeStore(
-                "InputArguments",
-                runtimeAttrs,
-                () -> runtimeBean.getInputArguments().toArray(new String[0]));
-        safeStore("LibraryPath", runtimeAttrs, runtimeBean::getLibraryPath);
-        safeStore("ManagementSpecVersion", runtimeAttrs, runtimeBean::getManagementSpecVersion);
-        safeStore("Name", runtimeAttrs, runtimeBean::getName);
-        safeStore("SpecName", runtimeAttrs, runtimeBean::getSpecName);
-        safeStore("SpecVersion", runtimeAttrs, runtimeBean::getSpecVersion);
-        safeStore("SystemProperties", runtimeAttrs, runtimeBean::getSystemProperties);
-        safeStore("StartTime", runtimeAttrs, runtimeBean::getStartTime);
-        safeStore("Uptime", runtimeAttrs, runtimeBean::getUptime);
-        safeStore("VmName", runtimeAttrs, runtimeBean::getVmName);
-        safeStore("VmVendor", runtimeAttrs, runtimeBean::getVmVendor);
-        safeStore("VmVersion", runtimeAttrs, runtimeBean::getVmVersion);
-        safeStore("BootClassPathSupported", runtimeAttrs, runtimeBean::isBootClassPathSupported);
-        RuntimeMetrics runtime = new RuntimeMetrics(runtimeAttrs);
-
-        MemoryMXBean memoryBean = ManagementFactory.getMemoryMXBean();
-        Map<String, Object> memoryAttrs = new HashMap<>();
-        safeStore("HeapMemoryUsage", memoryAttrs, memoryBean::getHeapMemoryUsage);
-        safeStore("NonHeapMemoryUsage", memoryAttrs, memoryBean::getNonHeapMemoryUsage);
-        safeStore(
-                "ObjectPendingFinalizationCount",
-                memoryAttrs,
-                memoryBean::getObjectPendingFinalizationCount);
-        // TODO are these inferred calculations correct (ie do they match what the bean attributes
-        // say)?
-        safeStore(
-                "FreeHeapMemory",
-                memoryAttrs,
-                () ->
-                        memoryBean.getHeapMemoryUsage().getCommitted()
-                                - memoryBean.getHeapMemoryUsage().getUsed());
-        safeStore(
-                "FreeNonHeapMemory",
-                memoryAttrs,
-                () ->
-                        memoryBean.getNonHeapMemoryUsage().getCommitted()
-                                - memoryBean.getNonHeapMemoryUsage().getUsed());
-        safeStore(
-                "HeapMemoryUsagePercent",
-                memoryAttrs,
-                () ->
-                        ((double) memoryBean.getHeapMemoryUsage().getUsed()
-                                / (double) memoryBean.getHeapMemoryUsage().getCommitted()));
-        safeStore("Verbose", memoryAttrs, memoryBean::isVerbose);
-        MemoryMetrics memory = new MemoryMetrics(memoryAttrs);
-
-        ThreadMXBean threadBean = ManagementFactory.getThreadMXBean();
-        Map<String, Object> threadAttrs = new HashMap<>();
-        safeStore("AllThreadIds", threadAttrs, threadBean::getAllThreadIds);
-        safeStore("CurrentThreadCpuTime", threadAttrs, threadBean::getCurrentThreadCpuTime);
-        safeStore("CurrentThreadUserTime", threadAttrs, threadBean::getCurrentThreadUserTime);
-        safeStore("DaemonThreadCount", threadAttrs, threadBean::getDaemonThreadCount);
-        safeStore("PeakThreadCount", threadAttrs, threadBean::getPeakThreadCount);
-        safeStore("ThreadCount", threadAttrs, threadBean::getThreadCount);
-        safeStore("TotalStartedThreadCount", threadAttrs, threadBean::getTotalStartedThreadCount);
-        safeStore(
-                "CurrentThreadCpuTimeSupported",
-                threadAttrs,
-                threadBean::isCurrentThreadCpuTimeSupported);
-        safeStore(
-                "ObjectMonitorUsageSupported",
-                threadAttrs,
-                threadBean::isObjectMonitorUsageSupported);
-        safeStore(
-                "SynchronizerUsageSupported",
-                threadAttrs,
-                threadBean::isSynchronizerUsageSupported);
-        safeStore(
-                "ThreadContentionMonitoringEnabled",
-                threadAttrs,
-                threadBean::isThreadContentionMonitoringEnabled);
-        safeStore(
-                "ThreadContentionMonitoringSupported",
-                threadAttrs,
-                threadBean::isThreadContentionMonitoringSupported);
-        safeStore("ThreadCpuTimeEnabled", threadAttrs, threadBean::isThreadCpuTimeEnabled);
-        safeStore("ThreadCpuTimeSupported", threadAttrs, threadBean::isThreadCpuTimeSupported);
-        ThreadMetrics threads = new ThreadMetrics(threadAttrs);
-
-        OperatingSystemMXBean osBean = ManagementFactory.getOperatingSystemMXBean();
-        Map<String, Object> osAttrs = new HashMap<>();
-        safeStore("Arch", osAttrs, osBean::getArch);
-        safeStore("AvailableProcessors", osAttrs, osBean::getAvailableProcessors);
-        safeStore("Name", osAttrs, osBean::getName);
-        safeStore("SystemLoadAverage", osAttrs, osBean::getSystemLoadAverage);
-        safeStore("Version", osAttrs, osBean::getVersion);
-        if (osBean instanceof UnixOperatingSystemMXBean) {
-            UnixOperatingSystemMXBean unix = (UnixOperatingSystemMXBean) osBean;
-            safeStore("CommittedVirtualMemorySize", osAttrs, unix::getCommittedVirtualMemorySize);
-            safeStore("FreePhysicalMemorySize", osAttrs, unix::getFreePhysicalMemorySize);
-            safeStore("FreeSwapSpaceSize", osAttrs, unix::getFreeSwapSpaceSize);
-            safeStore("ProcessCpuLoad", osAttrs, unix::getProcessCpuLoad);
-            safeStore("ProcessCpuTime", osAttrs, unix::getProcessCpuTime);
-            safeStore("SystemCpuLoad", osAttrs, unix::getSystemCpuLoad);
-            safeStore("TotalPhysicalMemorySize", osAttrs, unix::getTotalPhysicalMemorySize);
-            safeStore("TotalSwapSpaceSize", osAttrs, unix::getTotalSwapSpaceSize);
-        }
-        OperatingSystemMetrics os = new OperatingSystemMetrics(osAttrs);
-
-        String jvmId = null;
-        try (ByteArrayOutputStream baos = new ByteArrayOutputStream(512);
-                DataOutputStream dos = new DataOutputStream(baos)) {
-            dos.writeUTF(runtime.getClassPath());
-            dos.writeUTF(runtime.getName());
-            dos.writeUTF(Arrays.toString(runtime.getInputArguments()));
-            dos.writeUTF(runtime.getLibraryPath());
-            dos.writeUTF(runtime.getVmVendor());
-            dos.writeUTF(runtime.getVmVersion());
-            dos.writeLong(runtime.getStartTime());
-            byte[] hash = DigestUtils.sha256(baos.toByteArray());
-            jvmId = new String(Base64.getUrlEncoder().encode(hash), StandardCharsets.UTF_8).trim();
-        } catch (IOException e) {
-            log.error("Could not compute own jvmId!", e);
-        }
-
-        return new MBeanMetrics(runtime, memory, threads, os, jvmId);
     }
 
     private class AgentAuthenticator extends BasicAuthenticator {
